Fixes #9248 [Search Term Group] - Reload only the part of cell and not whole view
